(PC-23404)[PRO] refactor: remove BannerLayout index file
