--- pro/src/ui-kit/Banners/BannerLayout/index.ts
@@ -1,1 +1,0 @@
-export { default } from './BannerLayout'
